requirements.txt
requests>=2.12.4
dnspython>=1.15.0